Allow customizable login transition. Fixes #22
bug fix

hound report fixes

fix app/controller/login export and make hound shutup about es6 modules

removed duplicate esnext jshint config
