+++ app/routes/logout.js
@@ -1,8 +1,11 @@
 import Ember from 'ember';
+import AuthenticatedMixin from 'ember-authenticate-me/mixins/routes/authenticated';
 
-export default Ember.Route.extend({
+export default Ember.Route.extend(AuthenticatedMixin, {
   beforeModel: function() {
     var self = this;
+
+    this._super(...arguments);
 
     this.get('session').close().then(function() {
       self.transitionTo('login');
