+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "ember-authenticate-me",
-  "version": "0.0.19",
+  "version": "0.1.0",
   "description": "The default blueprint for ember-cli addons.",
   "directories": {
     "doc": "doc",
